+++ sickbeard/providers/__init__.py
@@ -26,7 +26,7 @@ from sickbeard import logger
 from . import newznab, omgwtfnzbs, womble
 # torrent
 from . import alpharatio, beyondhd, bitmetv, bitsoup, btn, freshontv, funfile, gftracker, grabtheinfo, \
-    hdbits, hdspace, iptorrents, kat, morethan, pisexy, pretome, rarbg, scc, scenetime, shazbat, speedcd, \
+    hdbits, hdspace, iptorrents, morethan, pisexy, pretome, rarbg, scc, scenetime, shazbat, speedcd, \
     thepiratebay, torrentbytes, torrentday, torrenting, torrentleech, torrentshack, transmithe_net, tvchaosuk
 # anime
 from . import nyaatorrents, tokyotoshokan
@@ -45,7 +45,6 @@ __all__ = ['omgwtfnzbs',
            'hdbits',
            'hdspace',
            'iptorrents',
-           'kat',
            'morethan',
            'pisexy',
            'pretome',