+++ CHANGES.md
@@ -105,6 +105,7 @@
 * Add qBitTorrent to Search Settings/Torrent Search
 * Add "Test NZBGet" client to Search Settings/NZB Search/NZBGet
 * Change include x265 category when searching IPT provider
+* Change init.systemd to use python2 binary and recommended installation paths
 
 [develop changelog]
 * Change send nzb data to NZBGet for Anizb instead of url